requirements.txt
deepspeed==0.9.5
numpy==1.24.1
opencv_python_headless==4.8.1.78
peft==0.4.0
Pillow==10.3.0
PyYAML==6.0.1
scikit_learn==1.2.2
scipy==1.13.0
spacy==3.7.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.35.0
sentencepiece==0.2.0
python-dotenv
